added all specific interfaces

+++ FYPAPI/Models/Fyp.cs
@@ -4,14 +4,14 @@ namespace FYPAPI.Models
 {
     public class Fyp: IEquatable<Fyp>
     {
-        public short PK_tblFyp { get; set; }
+        public int PK_tblFyp { get; set; }
         public string title { get; set; }
         public string description { get; set; }
         public string category { get; set; }
         public string type { get; set; }
         public string otherRequirements { get; set; }
-        public short minStudents { get; set; }
-        public short maxStudents { get; set; }
+        public int minStudents { get; set; }
+        public int maxStudents { get; set; }
         public bool isAvailable { get; set; }
 
         public bool Equals(Fyp other)
